initial clean
moved configs into their own folders
moved lazy list/set into configprovider and cleaned the code
