+++ src/main/java/lycanitestweaks/mixin/lycanitestweaksmajor/beastiaryclient/IndexBeastiaryScreenMixin.java
@@ -6,6 +6,7 @@ import com.lycanitesmobs.client.gui.beastiary.IndexBeastiaryScreen;
 import lycanitestweaks.capability.IPlayerMobLevelCapability;
 import lycanitestweaks.capability.PlayerMobLevelCapability;
 import lycanitestweaks.handlers.ForgeConfigHandler;
+import lycanitestweaks.handlers.ForgeConfigProvider;
 import lycanitestweaks.handlers.config.PlayerMobLevelsConfig;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.player.EntityPlayer;
@@ -51,7 +52,7 @@ public abstract class IndexBeastiaryScreenMixin extends BeastiaryScreen {
             String text = I18n.format("gui.beastiary.index.mixin.category");
             this.getFontRenderer().drawString(text, xOffset, yOffset, 0xFFFFFF, true);
             yOffset += 4 + this.getFontRenderer().getWordWrappedHeight(text, this.colLeftWidth);
-            for(PlayerMobLevelsConfig.BonusCategory category : ForgeConfigHandler.getPmlBonusCategoryClientRenderOrder()){
+            for(PlayerMobLevelsConfig.BonusCategory category : ForgeConfigProvider.getPmlBonusCategoryClientRenderOrder()){
                 if(lycanitesTweaks$pmlBonusCateogories.containsKey(category)){
                     if(PlayerMobLevelsConfig.getPmlBonusCategorySoulgazer().contains(category))
                         this.drawTexture(new ResourceLocation(LycanitesMobs.modid, "textures/items/soulgazer.png"),xOffset - 20, yOffset - 4, 0, 1, 1, 16 ,16);
