+++ src/main/java/lycanitestweaks/mixin/lycanitestweaksminor/customnamespawning/EntityArisaurFlowersaurMixin.java
@@ -2,7 +2,7 @@ package lycanitestweaks.mixin.lycanitestweaksminor.customnamespawning;
 
 import com.lycanitesmobs.core.entity.AgeableCreatureEntity;
 import com.lycanitesmobs.core.entity.creature.EntityArisaur;
-import lycanitestweaks.handlers.ForgeConfigHandler;
+import lycanitestweaks.handlers.ForgeConfigProvider;
 import net.minecraft.entity.IEntityLivingData;
 import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.World;
@@ -19,7 +19,7 @@ public abstract class EntityArisaurFlowersaurMixin extends AgeableCreatureEntity
     @Unique
     @Override
     public IEntityLivingData onInitialSpawn(DifficultyInstance difficulty, IEntityLivingData livingData) {
-        if(this.getVariantIndex() == 0 && ForgeConfigHandler.getFlowersaurBiomes().contains(this.world.getBiome(this.getPosition()).getRegistryName())){
+        if(this.getVariantIndex() == 0 && ForgeConfigProvider.getFlowersaurBiomes().contains(this.world.getBiome(this.getPosition()).getRegistryName())){
             this.setCustomNameTag("Flowersaur");
         }
         return super.onInitialSpawn(difficulty, livingData);
